requirements.txt
ebooklib>=0.18
python-docx>=1.0
lxml>=4.9
Pillow>=10.0
pytesseract>=0.3.10
pyinstaller>=6.0
